chore(swaps): consolidate swap utils to a single file
