+++ ui/app/pages/swap/swap-content/swap-gas-customize/swap-gas-customize.component.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { decimalToHex } from '../swap.utils'
+import { decimalToHex } from '../../swap.utils'
 
 export default class CustomizeGasButton extends Component {
